Make sure that runToFinish() returns after VM dies
There is a gap between vm.stop() call to complete VM shutdown. So it
could be the case that runToFinish() returns before VM dies, which may
cause confusion. To ensure that, mExecutorService.shutdown() is now only
called from onDied() callback. This also makes the test robust by
implicitly checking whether onDied() is called or not.

Bug: 238692795
Test: atest MicrodroidTests
Change-Id: I4204ff443f3415c6795598c325e03b8bd8fc1f25