Add android_module_lib_stubs_current to stub_only_libs
The framework-virtualization belongs to an non-updatable APEX, hence it
should be able to compile against unstable hidden platform APIs. In
order to achieve this we specify the sdk_version: "core_platform" and
add a the framework lib to the impl_only_libs of the
framework-virtualization.

However, this setup breaks the generation of stub srcs: it relies on the
sdk_version of the java_sdk_library, which in our case doesn't provide
the symbols from the Android SDK. To work around this problem we add the
android_module_lib_stubs_current to the stub_only_libs.

The stub_only_libs get propagate to the stub sources invocation in the
another change in the same topic.

Bug: 243512044
Test: builds
Change-Id: Ied0bfaf653edc2d16564b1f1b551d0624be1b407