More robust parsing of /proc/meminfo
The test assumed that an entry (ex: foo: 100 kB) in /proc/meminfo
doesn't have a tab character. That assumption was broken when the kernel
starts to emit an entry with a tab chracter. Specifically, the following
entry was added with a tab character(s) in front of the number.

SecPageTables:         0 kB

Fixing this by matching a line with a regex.

Bug: N/A
Test: run com.android.microdroid.test.MicrodroidHostTests#testMicrodroidRamUsage
with the updated kernel

Change-Id: I6061c0adb382e2513bf2abbfea5c9219800f55f8