Remove onRamdump callback
Ramdumps are still written to tombstoned, but we remove the callback
to the client.

Remove the option from vm that allows the ramdump to be written elsewhere.

I could have left the callback from VS and just ignored it in our
javalib, but adding extra work to all clients just to support the
niche vm usecase felt wrong.

Also fix a type.

Bug: 262538986
Bug: 261037705
Test: atest MicrodroidTests
Change-Id: If60ac5033a8b077fc35b499e056461782b5027fe