Drop the VM reference when stopped
A VM may exit even without a call to stop() or close(). When we notice
that, drop our reference to the binder interface (to allow GC to kick
in abd VS to exit) and drop our memory pressure callback.

This is mostly good housekeeping, but it also makes sure we don't end
up registering the callback multiple times if we call run() again on
the same VM without an intervening stop().

Bug: 238931615
Test: atest MicrodroidTests
Change-Id: If4c75ca11adf4239e4337405c24ff5db08707e7a