[avb_footer] Include initrd_* hashes as hash descriptors in kernel footer
This cl also removes the old initrd_* digests included as prop of kernel
hash descriptor.
The new footer will have 3 hash descriptors for kernel, initrd_normal
and initrd_debug.

This is the reland of the change aosp/2375868. The build of initrd on
linux-x86 that has incorrect format (b/264940248) is disabled in this
cl.

Bug: 260821553
Test: m microdroid_kernel_signed and inspect the output using `avbtool
info_image --image <output>`

Change-Id: I401e476f34dcd0ae0dd07e50d8fe19e3f2533d9a