CTS: Encrypted storage is inaccessible to diff VM
Test that encrypted storage of 1 pVM should not be accessible by a VM
with different identity.

Well not really.. The storage image of pvm1 (atleast after first boot)
contains encrypted filesystem which should mean nothing more than random
garbage when a different VM tries to access it. encryptedstore will fail
to 'mount' this filesystem.

This is equivalent to testing that sealing key derived for a different
instance image is different.

Test: #encryptedStorageIsInaccessibleToDiffVm
Bug: 259233794
Change-Id: I54cc84d7e1a5949879dd9bc7268d3a0e801c7649