pvmfw: refactor: Regroup config data slicing
As we need the slices covering the entries of the configuration data to
be mutable (e.g. to zero the BCC's CDIs or apply a DTBO), unify the API
so that both slices are returned through one call, to please the borrow
checked when the slices are used at the same time i.e. to avoid

    error[E0499]: cannot borrow `appended` as mutable more than once at a time

Note: No functional change intended.

Bug: 249054080
Bug: 268265510
Test: atest MicrodroidHostTests
Change-Id: I53f789add85f3f1b602a7b91cce3e7dd78727eef