Optimize identical configs
While writing a demo app I realized it makes sense to call
vm.setConfig immediately after vmm.getOrCreate - so if the VM already
existed but had a different config, we update it to what we want, or
find out it's incompatible.

And if in fact we just created the VM it would make sense for
setConfig to do minimal work, since old and new config will be the
same object.

Bug: 243512240
Test: atest MicrodroidTests
Change-Id: I0732eec3cd68ef9ad8d4e7b143ce49c46096f82b