Do not assume there is only one crosvm process
Our benchmarks search for crosvm's PID by running 'ps -a | grep crosvm'.
This may work now but will fail if/when then system runs VMs
independently of the test process.
Add methods to find crosvm as the child process of virtmgr, which in
turn is the child process of the current PID.

Test: atest MicrodroidBenchmarks
Change-Id: I0dcd1f3aba6f69becccb8a1573ea71046e932de0