pvmfw: apply_debug_policy: Improve failure path
As the function already makes an internal backup of the debug_policy,
turn the parameter into a shareable reference and pass the backup to
fdt_apply_overlay instead of the original DTBO. This removes the need to
do any work to restore the overlay on failure.

Avoid confusing logs such as

    [ERROR] Failed to apply debug policy: The requested node or property does not exist. Recovering...
    [INFO] Debug policy applied.

by returning a bool letting the caller know if DP was properly applied.

Report failure to apply DP as warnings instead of info! or error!.

Bug: 280425124
Test: atest MicrodroidHostTests
Change-Id: Ib5979b07c7978edd94dd90537faefebd25620b19
Merged-In: Ib5979b07c7978edd94dd90537faefebd25620b19