Don't forward console_in to both /dev/ttyS0 and /dev/hvc0
Previously, console_in was forwarded to both /dev/ttyS0 and /dev/hvc0.
This was wrong because data typed into console_in was sometimes consumed
by /dev/ttyS0, and sometimes by /dev/hvc0. This was non-deterministic
and also has caused test flakiness.

Fixing this by not sending the stream to /dev/ttyS0 which is for early
kernel outputs before virtio-console driver is loaded. We don't need the
serial device to be bi-directional.

Bug: 289732678
Test: run MicrodroidTestApp multiple times
Test: run Microdroid, type in something. check /dev/ttyS0 in the
Microdroid; it shows nothing. (Previously, it showed some partial
characters typed in.)

Change-Id: Ia8982fc7fbc917edb53de4e19bd1e81fa68a4fbf