vmbase: Use heap for non-protected SHARED_POOL
On non-protected platforms, buffers shared with devices (i.e. the host)
may be located on the heap as there are no security constraints given
that the host already has full access to guest memory. As the PCI/VirtIO
drivers expect SHARED_POOL to have been initialized and as the
init_{dynamic,static}_shared_pool() functions aren't appropriate for
non-protected VMs, provide a third method allowing clients to signal
that they want SHARED_POOL to be initialized to simply wrap the heap.

Use this mechanism in Rialto.

Test: atest DebugPolicyHostTests#testNoAdbInDebugPolicy_withDebugLevelNone_boots
Test: atest rialto_test vmbase_example.integration_test
Change-Id: I7098c8dfd3a5f2f447763f5f43ba880f2d8b008f