Don't run non-protected tests if not supported
We have various tests that use a non-protected VM. However, support
for those is not mandatory, and currently they fail on systems which
only support protected VMs.

This CL just prevents those tests from running if non-protected VMs
are not supported. This prevents spurous failures.

A better approach would be to run the tests on protected VMs instead;
that's in a follow-up CL.

To test this I modified ITestDevice to report no support for
non-protected VMs, and also modified VirtualizationManager to fail a
request for one. That allowed me to reproduce the failures and verify
that this gets rid of them.

Bug: 290727698
Test: atest MicrodroidHostTestCases

Change-Id: If748cf16f1ebcd8fec66595d3f172eca37d1648e