[vmbase] Adjust refill layout size in memory allocation
This cl adjusts the refill layout size to the next power of two
when allocating memory. This is necessary because the actual size
of the memory allocated in the subsequent call `alloc_aligned` is
the max of the next power of two and the alignment.

This is clear based on the implementation of `alloc_aligned`
but is not mentioned in its documentation[1].

Without this adjustment, the `alloc_aligned` call will fail if
given a layout whose size is not a power of two.

[1] https://docs.rs/buddy_system_allocator/latest/buddy_system_allocator/struct.FrameAllocator.html#method.alloc_aligned

Bug: 274441673
Test: atest MicrodroidTests
Change-Id: I37d7a2805aa6788d22c2cd5bfc8c701fad1f007d