Don't test unspecified behavior
VirtualMachineConfig#isCompatibleWith is defined to reject some
differences, and allow others, but in many cases doesn't specify what
the implementation should do.

Remove the test for one of the unspecified cases, since there are
valid implementations that treat it in different ways.

This CL should not merge to main as the code and test there has
changed; the Merged-In below refers to the commit that changed it.

Bug: 323448288
Bug: 321184334
Test: atest MicrodroidTests#compatibleConfigTests
Change-Id: Ib7b33c8e313ff7ae5c9616a6d4d67ffb5d57b174
Merged-In: I384360b0d19013b12da338af95cbc5ae104fa983