More lenient config descriptor parsing
Some DICE nodes simply have a hash in the config descriptor field; so
if parsing as the expected CBOR fails, we should ignore that.

Some nodes also have no name field, so we need to cope with that too.

And the SubComponents key in the config descriptor map is not
reserved, so it may appear in any DICE node with arbitrary content. So
we defer parsing the subcomponents until we've figured out which node
describes the Microdroid payload.

Bug: 338745127
Test: atest VmAttestationTestApp
Change-Id: I70169fc77170e2286d93246b079d3036d544b776