VS: Connect to Secretkeeper lazily
Rather than attempting to connect to Sk immediately if we believe it's
present, wait until we need it (for maintenance, or when
creating/deleting/updating a VM secret).

This avoids an issue where we fail to connect when Isolated
Compilation runs a VM during boot, where Secretkeeper is not yet
started.

Slightly gratuitously, split up some long imports to allow
auto-reformatting to work.

Bug: 331417880
Test: atest MicrodroidTests
Test: atest virtualizationservice_test
Test: Manually run CompOS, check no delay
Change-Id: I3db2070417bb1911fd8349e4bef9a420144ac245