New DICE chain requirements apply from VSR 15
For vendor API levels before 202404 there is no requirement that the
DICE chain for a VM be properly rooted, and so it is legitimate for
their to be no DICE entry preceding that for pvmwfw. Adjust the test
accordingly.

While I'm here, extract a common helper for reading the vendor API
level.

Bug: 341740108
Test: atest MicrodroidTestApp
Change-Id: Ic8c312e09acf30f92d3a868b4636b9d4dc520db9