pvmfw: Make pvmfw_embedded_key a proper PEM file
Allow pvmfw_embedded_key to be used for signing (e.g. Microdroid) by
making it hold the whole PEM, not just the binary pubkey, and derive the
pubkey locally from it, for embedding in libpvmfw_embedded_key.

Make it a filegroup as it should have been from the start given that
the key file is not part of the shipping images. Note that we previously
had to use a prebuilt_etc so that Make could obtain the file through
module-target-built-files but now that the pubkey is derived using a
genrule, it is seen by module-target-built-files without playing tricks.

Bug: 342549834
Test: m out/target/product/${PROD}/pvmfw_embedded.avbpubkey
Change-Id: I9a6602b727f719351292597b5b59b3e65936b800