Serial devices follows the legacy order
crosvm uses the same legacy order (3f8, 2f8, 3e8, 2e8) across x86 and
arm64. crosvm once tried to change that for non-x86 architecture
(because it doesn't have to), but that didn't land because there were
existing use case where the legacy order is expected.

This change makes sure that the serial device order is not changed by
the rewrite of the device tree by pvmfw.

Bug: 330659544
Test: run microdroid in pvm and check the kernel log
Change-Id: I3cfecc882148d0d1edd1a699123a0d26cee9ce95